requirements.txt
typer
fastapi
uvicorn
pydantic
httpx
colorama
pyyaml
tqdm
